// src/app/models/entrada/entrada-produto.ts
export class EntradaProduto {
    area: number = 0;
    cdBeneficiamento: number | undefined;
    cdCliente: number | undefined;
    cdProduto: number | undefined;
    data: string | undefined;
    id: number | undefined;
    nomeBeneficiamento: string = "";
    nomeCliente: string = "";
    nomeProduto: string = "";
    quantidade: number = 0;
    valorTotal: number = 0;
    valor: number = 0;
}